Support new versions of Thunder.
As maintainer of plugin I want to use it with any version of Thunder e.g. 5.0, master branch.
This repo is incopatible with new versions of Thunder due to bugs.
Fix CMakeLists.txt to use namespace options.
Fix CMakeLists.txt to not use option PROJECT_DIR as Thunder does not recongize it.
Fix IDRM.h to not redefine IShell as this causes build errors.